Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
marshmallow
marshmallow-sqlalchemy
flask-jwt-extended
python-dotenv

# dev
black
ruff
pytest
